+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -20,6 +20,9 @@ mode.
     - `lib/webdriver.WebElementPromise`
 * The `testing/index` module no longer wraps the promise manager
 * Removed `remote/index.DriverService.prototype.stop()` (use `#kill()` instead)
+* Removed the `firefox.Binary` class. Custom binaries can still be selected
+  using `firefox.Options#setBinary()`. Likewise, custom binary arguments can be
+  specified with `firefox.Options#addArguments()`.
 * Removed the `lib/actions` module
 * Removed the `phantomjs` module
 * Removed the 'opera' module